moved structre definitions from rosflight_structs.h into board.h, deleted rosflight_structs.h and updated files that included it to board.h
