+++ include/interface/board.h
@@ -36,10 +36,130 @@
 #include <cstddef>
 #include <cstdint>
 
-#include "rosflight_structs.h"
+#include "board.h"
 
 namespace rosflight_firmware
 {
+
+typedef struct //__attribute__((__packed__))
+{
+  uint64_t timestamp; // us, time of data read complete
+  uint64_t complete;  //
+  uint16_t status;    // device dependent
+} PacketHeader;
+
+typedef struct //__attribute__((__packed__))
+{
+  PacketHeader header;
+  float voltage;
+  float current;
+  float temperature; // STM32 temperature, not batter temperature
+} BatteryStruct;
+
+typedef struct //__attribute__((__packed__))
+{
+  PacketHeader header;
+  float accel[3];    // rad/s
+  float gyro[3];     // rad/s
+  float temperature; // K
+} ImuStruct;
+
+typedef struct //__attribute__((__packed__))
+{
+  PacketHeader header;
+  float pressure;    // Pa
+  float temperature; // K
+  union
+  {
+    float altitude;
+    float ias; //speed;
+  };
+} PressureStruct;
+
+enum class SensorRangeType // c.f., ROSFLIGHT_RANGE_TYPE
+{
+  ROSFLIGHT_RANGE_SONAR = 0, /*  | */
+  ROSFLIGHT_RANGE_LIDAR = 1, /*  | */
+  END = 2,                   /*  | */
+};
+
+typedef struct //__attribute__((__packed__))
+{
+  PacketHeader header;
+  float range;          // m
+  float min_range;      // m
+  float max_range;      // m
+  SensorRangeType type; // ROSFLIGHT_RANGE_SONAR, ROSFLIGHT_RANGE_SONAR
+} RangeStruct;
+
+typedef struct //__attribute__((packed))
+{
+  PacketHeader header; //
+  float flux[3];       // T, magnetic flux density
+  float temperature;   // K
+} MagStruct;
+
+enum class GNSSFixType // quality from GGA
+{
+  GNSS_FIX_TYPE_NO_FIX = 0,
+  GNSS_FIX_TYPE_DEAD_RECKONING_ONLY = 1,
+  GNSS_FIX_TYPE_2D_FIX = 2,
+  GNSS_FIX_TYPE_3D_FIX = 3,
+  GNSS_FIX_TYPE_GNSS_PLUS_DEAD_RECKONING = 4,
+  GNSS_FIX_TYPE_TIME_FIX_ONLY = 5,
+  GNSS_FIX_RTK_FLOAT = 6,
+  GNSS_FIX_RTK_FIXED = 7,
+  END = 8
+};
+
+typedef struct //__attribute__((__packed__))
+{
+  PacketHeader header;
+  uint64_t pps;           // most recent pps timestamp (us)
+  int64_t unix_seconds;   // Unix time, in seconds
+  int32_t unix_nanos;
+  uint8_t fix_type;
+  uint8_t num_sat;
+  double lon;
+  double lat;
+  float height_msl;
+  float vel_n;
+  float vel_e;
+  float vel_d;
+  float h_acc;
+  float v_acc;
+  float speed_accy;
+} GnssStruct;
+
+typedef struct
+{
+  bool imu;
+  bool gnss;
+  bool baro;
+  bool mag;
+  bool diff_pressure;
+  bool sonar;
+  bool battery;
+} got_flags;
+
+// 16 analog + 8 digital MUST BE > 14 (Mavlink message size is hardware to 14)
+#define RC_STRUCT_CHANNELS 24
+typedef struct //__attribute__((packed))
+{
+  PacketHeader header;
+  uint8_t nChan;
+  float chan[RC_STRUCT_CHANNELS];
+  bool frameLost;
+  bool failsafeActivated;
+} RcStruct;
+
+typedef struct //__attribute__((__packed__))
+{
+  PacketHeader header;
+  float q[4]; // quaternions
+  float rate[3];
+} AttitudeStruct;
+
 class Board
 {
 public:
